Create training_model.py
This file is responsible for training our Neural Network (defined in model.py) . At first we will iterate over the intents file to analyze the words, labels and tags and then create all the necessary datasets along with the hyper-parameter specifications required such as batch size learning rate, number of epochs etc. and finally it saves them to a .PTH data file so that it can be loaded on our chat interface later on.